Rewrite LWRP as a Custom Resource
